Enum gauthz::Scope

pub enum Scope {
    Activity,
    AdexchangeBuyer,
    AdexchangeSeller,
    AdexchangeSellerReadonly,
    AdminDatatransfer,
    AdminDatatransferReadonly,
    AdminDirectoryCustomer,
    AdminDirectoryCustomerReadonly,
    AdminDirectoryDeviceChromeos,
    AdminDirectoryDeviceChromeosReadonly,
    AdminDirectoryDeviceMobile,
    AdminDirectoryDeviceMobileAction,
    AdminDirectoryDeviceMobileReadonly,
    AdminDirectoryDomain,
    AdminDirectoryDomainReadonly,
    AdminDirectoryGroup,
    AdminDirectoryGroupMember,
    AdminDirectoryGroupMemberReadonly,
    AdminDirectoryGroupReadonly,
    AdminDirectoryNotifications,
    AdminDirectoryOrgunit,
    AdminDirectoryOrgunitReadonly,
    AdminDirectoryResourceCalendar,
    AdminDirectoryResourceCalendarReadonly,
    AdminDirectoryRolemanagement,
    AdminDirectoryRolemanagementReadonly,
    AdminDirectoryUser,
    AdminDirectoryUserAlias,
    AdminDirectoryUserAliasReadonly,
    AdminDirectoryUserReadonly,
    AdminDirectoryUserSecurity,
    AdminDirectoryUserschema,
    AdminDirectoryUserschemaReadonly,
    AdminReportsAuditReadonly,
    AdminReportsUsageReadonly,
    Adsense,
    AdsenseReadonly,
    Adsensehost,
    Analytics,
    AnalyticsEdit,
    AnalyticsManageUsers,
    AnalyticsManageUsersReadonly,
    AnalyticsProvision,
    AnalyticsReadonly,
    Androidenterprise,
    Androidpublisher,
    AppengineAdmin,
    AppsGroupsMigration,
    AppsGroupsSettings,
    AppsLicensing,
    AppsOrder,
    AppsOrderReadonly,
    Appstate,
    Bigquery,
    BigqueryInsertdata,
    Blogger,
    BloggerReadonly,
    Books,
    Calendar,
    CalendarFeeds,
    CalendarReadonly,
    ClassroomCourses,
    ClassroomCoursesReadonly,
    ClassroomCourseworkMe,
    ClassroomCourseworkMeReadonly,
    ClassroomCourseworkStudents,
    ClassroomCourseworkStudentsReadonly,
    ClassroomGuardianlinksMeReadonly,
    ClassroomGuardianlinksStudents,
    ClassroomGuardianlinksStudentsReadonly,
    ClassroomProfileEmails,
    ClassroomProfilePhotos,
    ClassroomRosters,
    ClassroomRostersReadonly,
    ClassroomStudentSubmissionsMeReadonly,
    ClassroomStudentSubmissionsStudentsReadonly,
    CloudDebugger,
    CloudLanguage,
    CloudPlatform,
    CloudPlatformReadOnly,
    CloudTranslation,
    CloudUseraccounts,
    CloudUseraccountsReadonly,
    Cloudruntimeconfig,
    Compute,
    ComputeReadonly,
    Contacts,
    ContactsReadonly,
    Content,
    Datastore,
    Ddmconversions,
    DevstorageFullControl,
    DevstorageReadOnly,
    DevstorageReadWrite,
    Dfareporting,
    Dfatrafficking,
    Doubleclickbidmanager,
    Doubleclicksearch,
    Drive,
    DriveAppdata,
    DriveFile,
    DriveMetadata,
    DriveMetadataReadonly,
    DrivePhotosReadonly,
    DriveReadonly,
    DriveScripts,
    Firebase,
    FirebaseReadonly,
    FitnessActivityRead,
    FitnessActivityWrite,
    FitnessBloodGlucoseRead,
    FitnessBloodGlucoseWrite,
    FitnessBloodPressureRead,
    FitnessBloodPressureWrite,
    FitnessBodyRead,
    FitnessBodyTemperatureRead,
    FitnessBodyTemperatureWrite,
    FitnessBodyWrite,
    FitnessLocationRead,
    FitnessLocationWrite,
    FitnessNutritionRead,
    FitnessNutritionWrite,
    FitnessOxygenSaturationRead,
    FitnessOxygenSaturationWrite,
    FitnessReproductiveHealthRead,
    FitnessReproductiveHealthWrite,
    Forms,
    FormsCurrentonly,
    Fusiontables,
    FusiontablesReadonly,
    Games,
    Genomics,
    GenomicsReadonly,
    GlassLocation,
    GlassTimeline,
    Gmail,
    GmailCompose,
    GmailInsert,
    GmailLabels,
    GmailMetadata,
    GmailModify,
    GmailReadonly,
    GmailSend,
    GmailSettingsBasic,
    GmailSettingsSharing,
    Groups,
    LoggingAdmin,
    LoggingRead,
    LoggingWrite,
    M8Feeds,
    Manufacturercenter,
    Monitoring,
    MonitoringRead,
    MonitoringWrite,
    NdevClouddnsReadonly,
    NdevClouddnsReadwrite,
    NdevCloudman,
    NdevCloudmanReadonly,
    PlaymoviesPartnerReadonly,
    PlusCirclesRead,
    PlusCirclesWrite,
    PlusLogin,
    PlusMe,
    PlusMediaUpload,
    PlusProfilesRead,
    PlusStreamRead,
    PlusStreamWrite,
    Prediction,
    Presentations,
    PresentationsReadonly,
    Pubsub,
    Replicapool,
    ReplicapoolReadonly,
    ServiceManagement,
    ServiceManagementReadonly,
    Servicecontrol,
    Siteverification,
    SiteverificationVerifyOnly,
    SourceReadOnly,
    SourceReadWrite,
    SpannerAdmin,
    SpannerData,
    Spreadsheets,
    SpreadsheetsReadonly,
    SqlserviceAdmin,
    TagmanagerDeleteContainers,
    TagmanagerEditContainers,
    TagmanagerEditContainerversions,
    TagmanagerManageAccounts,
    TagmanagerManageUsers,
    TagmanagerPublish,
    TagmanagerReadonly,
    Taskqueue,
    TaskqueueConsumer,
    Tasks,
    TasksReadonly,
    Urlshortener,
    UserAddressesRead,
    UserBirthdayRead,
    UserEmailsRead,
    UserPhonenumbersRead,
    UserinfoEmail,
    UserinfoProfile,
    UserlocationBeaconRegistry,
    Webmasters,
    WebmastersReadonly,
    XapiZoo,
    Youtube,
    YoutubeForceSsl,
    YoutubeReadonly,
    YoutubeUpload,
    Youtubepartner,
    YoutubepartnerChannelAudit,
    YtAnalyticsMonetaryReadonly,
    YtAnalyticsReadonly,
    Custom(&'static str),
}

API Scope values

For convenience, std::str::FromStr is defined for enabling cases where you have the url representing the scope and which to call url.parse() to resolve a Scope value

Variants
Activity

Scope for https://www.googleapis.com/auth/activity

AdexchangeBuyer

Scope for https://www.googleapis.com/auth/adexchange.buyer

AdexchangeSeller

Scope for https://www.googleapis.com/auth/adexchange.seller

AdexchangeSellerReadonly

Scope for https://www.googleapis.com/auth/adexchange.seller.readonly

AdminDatatransfer

Scope for https://www.googleapis.com/auth/admin.datatransfer

AdminDatatransferReadonly

Scope for https://www.googleapis.com/auth/admin.datatransfer.readonly

AdminDirectoryCustomer

Scope for https://www.googleapis.com/auth/admin.directory.customer

AdminDirectoryCustomerReadonly

Scope for https://www.googleapis.com/auth/admin.directory.customer.readonly

AdminDirectoryDeviceChromeos

Scope for https://www.googleapis.com/auth/admin.directory.device.chromeos

AdminDirectoryDeviceChromeosReadonly

Scope for https://www.googleapis.com/auth/admin.directory.device.chromeos.readonly

AdminDirectoryDeviceMobile

Scope for https://www.googleapis.com/auth/admin.directory.device.mobile

AdminDirectoryDeviceMobileAction

Scope for https://www.googleapis.com/auth/admin.directory.device.mobile.action

AdminDirectoryDeviceMobileReadonly

Scope for https://www.googleapis.com/auth/admin.directory.device.mobile.readonly

AdminDirectoryDomain

Scope for https://www.googleapis.com/auth/admin.directory.domain

AdminDirectoryDomainReadonly

Scope for https://www.googleapis.com/auth/admin.directory.domain.readonly

AdminDirectoryGroup

Scope for https://www.googleapis.com/auth/admin.directory.group

AdminDirectoryGroupMember

Scope for https://www.googleapis.com/auth/admin.directory.group.member

AdminDirectoryGroupMemberReadonly

Scope for https://www.googleapis.com/auth/admin.directory.group.member.readonly

AdminDirectoryGroupReadonly

Scope for https://www.googleapis.com/auth/admin.directory.group.readonly

AdminDirectoryNotifications

Scope for https://www.googleapis.com/auth/admin.directory.notifications

AdminDirectoryOrgunit

Scope for https://www.googleapis.com/auth/admin.directory.orgunit

AdminDirectoryOrgunitReadonly

Scope for https://www.googleapis.com/auth/admin.directory.orgunit.readonly

AdminDirectoryResourceCalendar

Scope for https://www.googleapis.com/auth/admin.directory.resource.calendar

AdminDirectoryResourceCalendarReadonly

Scope for https://www.googleapis.com/auth/admin.directory.resource.calendar.readonly

AdminDirectoryRolemanagement

Scope for https://www.googleapis.com/auth/admin.directory.rolemanagement

AdminDirectoryRolemanagementReadonly

Scope for https://www.googleapis.com/auth/admin.directory.rolemanagement.readonly

AdminDirectoryUser

Scope for https://www.googleapis.com/auth/admin.directory.user

AdminDirectoryUserAlias

Scope for https://www.googleapis.com/auth/admin.directory.user.alias

AdminDirectoryUserAliasReadonly

Scope for https://www.googleapis.com/auth/admin.directory.user.alias.readonly

AdminDirectoryUserReadonly

Scope for https://www.googleapis.com/auth/admin.directory.user.readonly

AdminDirectoryUserSecurity

Scope for https://www.googleapis.com/auth/admin.directory.user.security

AdminDirectoryUserschema

Scope for https://www.googleapis.com/auth/admin.directory.userschema

AdminDirectoryUserschemaReadonly

Scope for https://www.googleapis.com/auth/admin.directory.userschema.readonly

AdminReportsAuditReadonly

Scope for https://www.googleapis.com/auth/admin.reports.audit.readonly

AdminReportsUsageReadonly

Scope for https://www.googleapis.com/auth/admin.reports.usage.readonly

Adsense

Scope for https://www.googleapis.com/auth/adsense

AdsenseReadonly

Scope for https://www.googleapis.com/auth/adsense.readonly

Adsensehost

Scope for https://www.googleapis.com/auth/adsensehost

Analytics

Scope for https://www.googleapis.com/auth/analytics

AnalyticsEdit

Scope for https://www.googleapis.com/auth/analytics.edit

AnalyticsManageUsers

Scope for https://www.googleapis.com/auth/analytics.manage.users

AnalyticsManageUsersReadonly

Scope for https://www.googleapis.com/auth/analytics.manage.users.readonly

AnalyticsProvision

Scope for https://www.googleapis.com/auth/analytics.provision

AnalyticsReadonly

Scope for https://www.googleapis.com/auth/analytics.readonly

Androidenterprise

Scope for https://www.googleapis.com/auth/androidenterprise

Androidpublisher

Scope for https://www.googleapis.com/auth/androidpublisher

AppengineAdmin

Scope for https://www.googleapis.com/auth/appengine.admin

AppsGroupsMigration

Scope for https://www.googleapis.com/auth/apps.groups.migration

AppsGroupsSettings

Scope for https://www.googleapis.com/auth/apps.groups.settings

AppsLicensing

Scope for https://www.googleapis.com/auth/apps.licensing

AppsOrder

Scope for https://www.googleapis.com/auth/apps.order

AppsOrderReadonly

Scope for https://www.googleapis.com/auth/apps.order.readonly

Appstate

Scope for https://www.googleapis.com/auth/appstate

Bigquery

Scope for https://www.googleapis.com/auth/bigquery

BigqueryInsertdata

Scope for https://www.googleapis.com/auth/bigquery.insertdata

Blogger

Scope for https://www.googleapis.com/auth/blogger

BloggerReadonly

Scope for https://www.googleapis.com/auth/blogger.readonly

Books

Scope for https://www.googleapis.com/auth/books

Calendar

Scope for https://www.googleapis.com/auth/calendar

CalendarFeeds

Scope for https://www.google.com/calendar/feeds

CalendarReadonly

Scope for https://www.googleapis.com/auth/calendar.readonly

ClassroomCourses

Scope for https://www.googleapis.com/auth/classroom.courses

ClassroomCoursesReadonly

Scope for https://www.googleapis.com/auth/classroom.courses.readonly

ClassroomCourseworkMe

Scope for https://www.googleapis.com/auth/classroom.coursework.me

ClassroomCourseworkMeReadonly

Scope for https://www.googleapis.com/auth/classroom.coursework.me.readonly

ClassroomCourseworkStudents

Scope for https://www.googleapis.com/auth/classroom.coursework.students

ClassroomCourseworkStudentsReadonly

Scope for https://www.googleapis.com/auth/classroom.coursework.students.readonly

ClassroomGuardianlinksMeReadonly

Scope for https://www.googleapis.com/auth/classroom.guardianlinks.me.readonly

ClassroomGuardianlinksStudents

Scope for https://www.googleapis.com/auth/classroom.guardianlinks.students

ClassroomGuardianlinksStudentsReadonly

Scope for https://www.googleapis.com/auth/classroom.guardianlinks.students.readonly

ClassroomProfileEmails

Scope for https://www.googleapis.com/auth/classroom.profile.emails

ClassroomProfilePhotos

Scope for https://www.googleapis.com/auth/classroom.profile.photos

ClassroomRosters

Scope for https://www.googleapis.com/auth/classroom.rosters

ClassroomRostersReadonly

Scope for https://www.googleapis.com/auth/classroom.rosters.readonly

ClassroomStudentSubmissionsMeReadonly

Scope for https://www.googleapis.com/auth/classroom.student-submissions.me.readonly

ClassroomStudentSubmissionsStudentsReadonly

Scope for https://www.googleapis.com/auth/classroom.student-submissions.students.readonly

CloudDebugger

Scope for https://www.googleapis.com/auth/cloud_debugger

CloudLanguage

Scope for https://www.googleapis.com/auth/cloud-language

CloudPlatform

Scope for https://www.googleapis.com/auth/cloud-platform

CloudPlatformReadOnly

Scope for https://www.googleapis.com/auth/cloud-platform.read-only

CloudTranslation

Scope for https://www.googleapis.com/auth/cloud-translation

CloudUseraccounts

Scope for https://www.googleapis.com/auth/cloud.useraccounts

CloudUseraccountsReadonly

Scope for https://www.googleapis.com/auth/cloud.useraccounts.readonly

Cloudruntimeconfig

Scope for https://www.googleapis.com/auth/cloudruntimeconfig

Compute

Scope for https://www.googleapis.com/auth/compute

ComputeReadonly

Scope for https://www.googleapis.com/auth/compute.readonly

Contacts

Scope for https://www.googleapis.com/auth/contacts

ContactsReadonly

Scope for https://www.googleapis.com/auth/contacts.readonly

Content

Scope for https://www.googleapis.com/auth/content

Datastore

Scope for https://www.googleapis.com/auth/datastore

Ddmconversions

Scope for https://www.googleapis.com/auth/ddmconversions

DevstorageFullControl

Scope for https://www.googleapis.com/auth/devstorage.full_control

DevstorageReadOnly

Scope for https://www.googleapis.com/auth/devstorage.read_only

DevstorageReadWrite

Scope for https://www.googleapis.com/auth/devstorage.read_write

Dfareporting

Scope for https://www.googleapis.com/auth/dfareporting

Dfatrafficking

Scope for https://www.googleapis.com/auth/dfatrafficking

Doubleclickbidmanager

Scope for https://www.googleapis.com/auth/doubleclickbidmanager

Doubleclicksearch

Scope for https://www.googleapis.com/auth/doubleclicksearch

Drive

Scope for https://www.googleapis.com/auth/drive

DriveAppdata

Scope for https://www.googleapis.com/auth/drive.appdata

DriveFile

Scope for https://www.googleapis.com/auth/drive.file

DriveMetadata

Scope for https://www.googleapis.com/auth/drive.metadata

DriveMetadataReadonly

Scope for https://www.googleapis.com/auth/drive.metadata.readonly

DrivePhotosReadonly

Scope for https://www.googleapis.com/auth/drive.photos.readonly

DriveReadonly

Scope for https://www.googleapis.com/auth/drive.readonly

DriveScripts

Scope for https://www.googleapis.com/auth/drive.scripts

Firebase

Scope for https://www.googleapis.com/auth/firebase

FirebaseReadonly

Scope for https://www.googleapis.com/auth/firebase.readonly

FitnessActivityRead

Scope for https://www.googleapis.com/auth/fitness.activity.read

FitnessActivityWrite

Scope for https://www.googleapis.com/auth/fitness.activity.write

FitnessBloodGlucoseRead

Scope for https://www.googleapis.com/auth/fitness.blood_glucose.read

FitnessBloodGlucoseWrite

Scope for https://www.googleapis.com/auth/fitness.blood_glucose.write

FitnessBloodPressureRead

Scope for https://www.googleapis.com/auth/fitness.blood_pressure.read

FitnessBloodPressureWrite

Scope for https://www.googleapis.com/auth/fitness.blood_pressure.write

FitnessBodyRead

Scope for https://www.googleapis.com/auth/fitness.body.read

FitnessBodyTemperatureRead

Scope for https://www.googleapis.com/auth/fitness.body_temperature.read

FitnessBodyTemperatureWrite

Scope for https://www.googleapis.com/auth/fitness.body_temperature.write

FitnessBodyWrite

Scope for https://www.googleapis.com/auth/fitness.body.write

FitnessLocationRead

Scope for https://www.googleapis.com/auth/fitness.location.read

FitnessLocationWrite

Scope for https://www.googleapis.com/auth/fitness.location.write

FitnessNutritionRead

Scope for https://www.googleapis.com/auth/fitness.nutrition.read

FitnessNutritionWrite

Scope for https://www.googleapis.com/auth/fitness.nutrition.write

FitnessOxygenSaturationRead

Scope for https://www.googleapis.com/auth/fitness.oxygen_saturation.read

FitnessOxygenSaturationWrite

Scope for https://www.googleapis.com/auth/fitness.oxygen_saturation.write

FitnessReproductiveHealthRead

Scope for https://www.googleapis.com/auth/fitness.reproductive_health.read

FitnessReproductiveHealthWrite

Scope for https://www.googleapis.com/auth/fitness.reproductive_health.write

Forms

Scope for https://www.googleapis.com/auth/forms

FormsCurrentonly

Scope for https://www.googleapis.com/auth/forms.currentonly

Fusiontables

Scope for https://www.googleapis.com/auth/fusiontables

FusiontablesReadonly

Scope for https://www.googleapis.com/auth/fusiontables.readonly

Games

Scope for https://www.googleapis.com/auth/games

Genomics

Scope for https://www.googleapis.com/auth/genomics

GenomicsReadonly

Scope for https://www.googleapis.com/auth/genomics.readonly

GlassLocation

Scope for https://www.googleapis.com/auth/glass.location

GlassTimeline

Scope for https://www.googleapis.com/auth/glass.timeline

Gmail

Scope for https://mail.google.com/

GmailCompose

Scope for https://www.googleapis.com/auth/gmail.compose

GmailInsert

Scope for https://www.googleapis.com/auth/gmail.insert

GmailLabels

Scope for https://www.googleapis.com/auth/gmail.labels

GmailMetadata

Scope for https://www.googleapis.com/auth/gmail.metadata

GmailModify

Scope for https://www.googleapis.com/auth/gmail.modify

GmailReadonly

Scope for https://www.googleapis.com/auth/gmail.readonly

GmailSend

Scope for https://www.googleapis.com/auth/gmail.send

GmailSettingsBasic

Scope for https://www.googleapis.com/auth/gmail.settings.basic

GmailSettingsSharing

Scope for https://www.googleapis.com/auth/gmail.settings.sharing

Groups

Scope for https://www.googleapis.com/auth/groups

LoggingAdmin

Scope for https://www.googleapis.com/auth/logging.admin

LoggingRead

Scope for https://www.googleapis.com/auth/logging.read

LoggingWrite

Scope for https://www.googleapis.com/auth/logging.write

M8Feeds

Scope for https://www.google.com/m8/feeds

Manufacturercenter

Scope for https://www.googleapis.com/auth/manufacturercenter

Monitoring

Scope for https://www.googleapis.com/auth/monitoring

MonitoringRead

Scope for https://www.googleapis.com/auth/monitoring.read

MonitoringWrite

Scope for https://www.googleapis.com/auth/monitoring.write

NdevClouddnsReadonly

Scope for https://www.googleapis.com/auth/ndev.clouddns.readonly

NdevClouddnsReadwrite

Scope for https://www.googleapis.com/auth/ndev.clouddns.readwrite

NdevCloudman

Scope for https://www.googleapis.com/auth/ndev.cloudman

NdevCloudmanReadonly

Scope for https://www.googleapis.com/auth/ndev.cloudman.readonly

PlaymoviesPartnerReadonly

Scope for https://www.googleapis.com/auth/playmovies_partner.readonly

PlusCirclesRead

Scope for https://www.googleapis.com/auth/plus.circles.read

PlusCirclesWrite

Scope for https://www.googleapis.com/auth/plus.circles.write

PlusLogin

Scope for https://www.googleapis.com/auth/plus.login

PlusMe

Scope for https://www.googleapis.com/auth/plus.me

PlusMediaUpload

Scope for https://www.googleapis.com/auth/plus.media.upload

PlusProfilesRead

Scope for https://www.googleapis.com/auth/plus.profiles.read

PlusStreamRead

Scope for https://www.googleapis.com/auth/plus.stream.read

PlusStreamWrite

Scope for https://www.googleapis.com/auth/plus.stream.write

Prediction

Scope for https://www.googleapis.com/auth/prediction

Presentations

Scope for https://www.googleapis.com/auth/presentations

PresentationsReadonly

Scope for https://www.googleapis.com/auth/presentations.readonly

Pubsub

Scope for https://www.googleapis.com/auth/pubsub

Replicapool

Scope for https://www.googleapis.com/auth/replicapool

ReplicapoolReadonly

Scope for https://www.googleapis.com/auth/replicapool.readonly

ServiceManagement

Scope for https://www.googleapis.com/auth/service.management

ServiceManagementReadonly

Scope for https://www.googleapis.com/auth/service.management.readonly

Servicecontrol

Scope for https://www.googleapis.com/auth/servicecontrol

Siteverification

Scope for https://www.googleapis.com/auth/siteverification

SiteverificationVerifyOnly

Scope for https://www.googleapis.com/auth/siteverification.verify_only

SourceReadOnly

Scope for https://www.googleapis.com/auth/source.read_only

SourceReadWrite

Scope for https://www.googleapis.com/auth/source.read_write

SpannerAdmin

Scope for https://www.googleapis.com/auth/spanner.admin

SpannerData

Scope for https://www.googleapis.com/auth/spanner.data

Spreadsheets

Scope for https://www.googleapis.com/auth/spreadsheets

SpreadsheetsReadonly

Scope for https://www.googleapis.com/auth/spreadsheets.readonly

SqlserviceAdmin

Scope for https://www.googleapis.com/auth/sqlservice.admin

TagmanagerDeleteContainers

Scope for https://www.googleapis.com/auth/tagmanager.delete.containers

TagmanagerEditContainers

Scope for https://www.googleapis.com/auth/tagmanager.edit.containers

TagmanagerEditContainerversions

Scope for https://www.googleapis.com/auth/tagmanager.edit.containerversions

TagmanagerManageAccounts

Scope for https://www.googleapis.com/auth/tagmanager.manage.accounts

TagmanagerManageUsers

Scope for https://www.googleapis.com/auth/tagmanager.manage.users

TagmanagerPublish

Scope for https://www.googleapis.com/auth/tagmanager.publish

TagmanagerReadonly

Scope for https://www.googleapis.com/auth/tagmanager.readonly

Taskqueue

Scope for https://www.googleapis.com/auth/taskqueue

TaskqueueConsumer

Scope for https://www.googleapis.com/auth/taskqueue.consumer

Tasks

Scope for https://www.googleapis.com/auth/tasks

TasksReadonly

Scope for https://www.googleapis.com/auth/tasks.readonly

Urlshortener

Scope for https://www.googleapis.com/auth/urlshortener

UserAddressesRead

Scope for https://www.googleapis.com/auth/user.addresses.read

UserBirthdayRead

Scope for https://www.googleapis.com/auth/user.birthday.read

UserEmailsRead

Scope for https://www.googleapis.com/auth/user.emails.read

UserPhonenumbersRead

Scope for https://www.googleapis.com/auth/user.phonenumbers.read

UserinfoEmail

Scope for https://www.googleapis.com/auth/userinfo.email

UserinfoProfile

Scope for https://www.googleapis.com/auth/userinfo.profile

UserlocationBeaconRegistry

Scope for https://www.googleapis.com/auth/userlocation.beacon.registry

Webmasters

Scope for https://www.googleapis.com/auth/webmasters

WebmastersReadonly

Scope for https://www.googleapis.com/auth/webmasters.readonly

XapiZoo

Scope for https://www.googleapis.com/auth/xapi.zoo

Youtube

Scope for https://www.googleapis.com/auth/youtube

YoutubeForceSsl

Scope for https://www.googleapis.com/auth/youtube.force-ssl

YoutubeReadonly

Scope for https://www.googleapis.com/auth/youtube.readonly

YoutubeUpload

Scope for https://www.googleapis.com/auth/youtube.upload

Youtubepartner

Scope for https://www.googleapis.com/auth/youtubepartner

YoutubepartnerChannelAudit

Scope for https://www.googleapis.com/auth/youtubepartner-channel-audit

YtAnalyticsMonetaryReadonly

Scope for https://www.googleapis.com/auth/yt-analytics-monetary.readonly

YtAnalyticsReadonly

Scope for https://www.googleapis.com/auth/yt-analytics.readonly

Custom(&'static str)

A custom API scope

Methods

impl Scope

fn url(&self) -> &'static str

Return url assocaited with scope

fn scope(url: &str) -> Option<Scope>

Return Scope assocaited with url